+++ lib/src/models/mixins/messageable.dart
@@ -11,7 +11,7 @@ import '../subreddit.dart';
 
 /// A mixin containing functionality to send messages to other [Redditor]s or
 /// [Subreddit] moderators.
-abstract class MessageableMixin {
+mixin MessageableMixin {
   Reddit get reddit;
   String get displayName;